+++ src/components/placeCheckin/registerVisitRequest.ts
@@ -1,9 +1,9 @@
-import { RegisterCheckinPayload } from "@/app/api/visit/register/route";
+import { RegisterCheckinPayload } from "@/app/api/checkin/place/route";
 
 export async function requestRegisterVisit(
   payload: RegisterCheckinPayload,
 ): Promise<void> {
-  await fetch("/api/visit/register", {
+  await fetch("/api/checkin/place", {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
